Thomas Eichelberger - Destroyers 2 - Electric Bogaloo
not much gameplay to see, I mainly got the movement and firing down for the ships. everything in the criteria is there except for an array. not sure if this counts for the 5% bonus to things we havent done,  but I did use a switch and coroutine found in the Moving_Ship script